src/notification/requirements.txt
confluent_kafka
python-dotenv
